Add seeds.rb script to load pharmacies from CSV
Leads to some questions about the data we hold on these pharmacies, in
particular what uniquely defines a pharmacy.
At the minute the seeds script expects the pharmacy to be unique by
[:name, :address_1, :address_3], but this is not enforced in the
application nor on the DB.

Also the assignment of the first column to address_3 does not work
universally well.