Rebuild solr indices as part of deployment
Currently indexes are stored on ephemeral storage which will vanish once
the container is shut down.
Need to look at the implications for this if/when containers undergo an
unexpected shutdown. But this is a bigger task as I don't think we are
generally handling/recovering in these cases generally.